requirements.txt
argparse
numpy
pandas
matplotlib
seaborn
scikit-learn
scipy
numba
tqdm
plotly
openpyxl
python-dotenv